+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [0.8.1] - 2020-10-07
 
-### Changeed
+### Changed
 
 - Allow structs in cursor values.